Test moving it inside?
